Moving initialization and cleanup output and error handling inside of classes
Messages were being output to the command line from Spectrometer.m during initialization and cleanup. These messages pertained to object that were being initialized/cleaned up by Spectrometer.m. These messages should be part of the classes that are being initialized/cleaned up and should not be Spectrometer.m's responsibility to update on the status of object creation/deletion. The same thing was fixed for some error handling. Spectrometer.m was using some try/catch blocks when initializing objects. These try/catch blocks should be inside of the class so that the errors are caught by the class constructor. This ensures that error handling works the same for all instances of the class.