Created an interface for Polarizer and Waveplate stages
The Thorlabs_DCServo class was created as a superclass for Thorlabs_KCubeDCServo and Thorlabs_TCubeDCServo. The superclass handles most of the functionality of the subclasses. This is because the T-Cube and K-Cube work similarly. The only difference between them is the .NET assemblies used and the function called to initialize the .NET server.