Adding new methods for polarization controlled 2DIR and pump-probe
New methods were created to execute polarization controlled 2DIR and pump-probe experiments (Method_2d_quasiphase_polarization and Method_Pump_Probe_polarization). Each of these methods inherits everything from their parent classes. The only difference is that the Scan method is now wrapped with a Scan method that runs the Scan@Super in the parallel and then the perpendicular polarization. I also had to the global variable 'rotors' to the method constructor in all of the Method subclasses and I fixed the indentation of all of the classes and subclasses of Method.