Changing when DC servo object was added to array
The DC servo object was being added to the rotors array inside the try/catch block. This led to objects not being added to the array in the event of an error. This was not the desired behavior because the a simulation object is generated when an error occurs during construction. This should still be added to the rotors array. To fix this the addition of the object to the rotors array was moved outside of the try/catch block.