Fixed some data type issues and table variable declarations
There was some confusion between cell and double in the time point data entry table and double was settled on for the data type of the table. There was also some issues with variable names not being redefined when the table was reinitialized after adding and removing rows of the table.