Fixing table behavior
When a row was added or removed from the time point entry table it was not pulling the current values in the table into the temporary variables and was overwriting changes made in the table. This was fixed by first reading the table and storing the data in the temp vars and then adding or removing a row and updating the table.