Fixing how the Scan stopping variable is changed
Changing the was that the ScanIsStopping variable is altered. The way that it was set ScanIsStopping to false at the end of each Method.Scan(). This would cause the ScanIsStopping to be set to false in all submethods. This would then lead to the loop continuing to the next time/polarization state even though the stop button was pressed.