Moved Lab Archives insert entry template from FileSystem to labarchivesCallObj
The file system was inserting an entry template into the lab archives page every time the file system was started. To fix this problem the insert of the entry template was moved to the labarchivesCallObj class. Now when the labarchivesCallObj creates a new page it will automatically insert a new entry template into the new notebook page.