Changing the angles of waveplate and polarizer; Adding a check for ScanIsStopping flag
The parallel and perpendicular pumps should be similar in intensity. In order for this to occur, the waveplate angle for parallel pumps was determined to be 21 degrees experimentally. The polarizer will now go to 270 degrees instead of 90 degrees for perpendicular because there are scratches on one side of the polarizer. Finally, before executing the Method superclass it will check if the stop button has been pressed.