If the arrays in the table do not have an empty value in the last row it will now add one
The uitable was adding the cell arrays as arrays if all cells were filled with numbers and causing data type errors. If you add an empty cell at the end of the cell array it doesn't do this. So, now it checks to see if the cell array has an empty cell and if not it adds one.